requirements.txt
cycler==0.11.0
fonttools==4.32.0
imageio==2.19.1
kiwisolver==1.4.2
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
PyQt6-Qt6==6.2.4
PyQt6-sip==13.3.1
PySide6==6.2.4
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.0
shiboken6==6.2.4
six==1.16.0
tifffile==2022.5.4
